requirements.txt
uvicorn
fastapi
SQLAlchemy
alembic
email-validator
pydantic
pydantic-settings
psycopg
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
jwt
python-multipart
PyJWT
greenlet
